pipelines/air_quality/assets/ingest/requirements.txt
requests
google-cloud-bigquery
google-auth